+++ test/lib/docx/docxBuilder.test.ts
@@ -1,7 +1,7 @@
 import { Document } from 'docx';
 import { Flow } from '../../../src/types/metadata/flow';
 
-import testFlow from '../testFlow.json';
+import testFlow from '../../data/testFlow.json';
 import DocxBuilder from '../../../src/lib/docx/docxBuilder';
 import { ReadableProcessMetadataConverter } from '../../../src/lib/converter/metadataConverter';
